+++ documentation/changelog.md
@@ -57,6 +57,7 @@ description: List of all the changes and current versions.
 + Some Interface Work
 + Fix Marker Scale for Jobs
 + Make Creation Interfaces Return UID(s)
++ Added Custom Streamer Key(s) based on Dimension and Position
 ```
 
 ## 2.0.3